add caching registry, refactor sql registry
